+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: lseabra- <lseabra-@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/09 14:03:13 by lseabra-          #+#    #+#             */
-/*   Updated: 2025/06/27 13:41:31 by lseabra-         ###   ########.fr       */
+/*   Updated: 2025/06/27 14:01:01 by lseabra-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 typedef struct s_stack
 {
 	char	*name;
-	long	*arr;
+	int	*arr;
 	int		len;
 }			t_stack;
